// indexer/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OrderStatus = {
  Open: 'Open',
  Partial: 'Partial',
  filled: 'filled',
  Cancelled: 'Cancelled',
  Settled: 'Settled'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]